sorter and video
