+++ packages/perseus/src/index.ts
@@ -254,4 +254,7 @@ export type {
     TagsShape,
 } from "./multi-items/shape-types";
 export type {Path} from "./multi-items/trees";
-export type {RendererPromptJSON, WidgetPromptJSON} from "./prompt-types";
+export type {
+    RendererPromptJSON,
+    WidgetPromptJSON,
+} from "./widget-ai-utils/prompt-types";
