+++ packages/perseus/src/types.ts
@@ -7,7 +7,6 @@ import type {
     PerseusWidget,
     PerseusWidgetsMap,
 } from "./perseus-types";
-import type {WidgetPromptJSON} from "./prompt-types";
 import type {PerseusStrings} from "./strings";
 import type {SizeClass} from "./util/sizing-utils";
 import type {
@@ -16,6 +15,7 @@ import type {
     UserInputArray,
     UserInputMap,
 } from "./validation.types";
+import type {WidgetPromptJSON} from "./widget-ai-utils/prompt-types";
 import type {KeypadAPI} from "@khanacademy/math-input";
 import type {AnalyticsEventHandlerFn} from "@khanacademy/perseus-core";
 import type {LinterContextProps} from "@khanacademy/perseus-linter";
