+++ packages/perseus/src/widget-ai-utils/graded-group/graded-group-ai-utils.ts
@@ -1,4 +1,4 @@
-import type {RendererPromptJSON} from "../../prompt-types";
+import type {RendererPromptJSON} from "../prompt-types";
 
 export type GradedGroupPromptJSON = RendererPromptJSON & {
     type: "graded-group";
