+++ packages/perseus/src/widget-ai-utils/sorter/sorter-ai-utils.test.ts
@@ -1,8 +1,8 @@
-import {getPromptJSON} from "./prompt-utils";
+import {getPromptJSON} from "./sorter-ai-utils";
 
 import type {PerseusSorterUserInput} from "../../validation.types";
 
-describe("Sorter getPromptJSON", () => {
+describe("Sorter AI utils", () => {
     it("it returns JSON with the expected format and fields", () => {
         const userInput: PerseusSorterUserInput = {
             options: ["Pickles", "Tomato", "Onion", "Lettuce"],
